+++ packages/sitecore-jss-vue/src/components/Link.test.ts
@@ -76,6 +76,43 @@ describe('<Link />', () => {
     expect(rendered.text()).toBe(props.field.text);
   });
 
+  it('should render with provided children', () => {
+    const props = {
+      field: {
+        href: '/lorem',
+        text: '[ipsum]',
+      },
+    };
+    const rendered = mount(Link, {
+      props,
+      slots: {
+        default: ['<p>Custom description</p>'],
+      },
+    }).find('a');
+
+    expect(rendered.attributes().href).toBe(props.field.href);
+    expect(rendered.text()).toBe('Custom description');
+  });
+
+  it('should render link text with provided children', () => {
+    const props = {
+      field: {
+        href: '/lorem',
+        text: '[ipsum]',
+      },
+      showLinkTextWithChildrenPresent: true,
+    };
+    const rendered = mount(Link, {
+      props,
+      slots: {
+        default: ['<p>Custom description</p>'],
+      },
+    }).find('a');
+
+    expect(rendered.attributes().href).toBe(props.field.href);
+    expect(rendered.text()).toBe('[ipsum]Custom description');
+  });
+
   it('should render ee HTML', () => {
     const props = {
       field: {